Misc. refactoring

+++ BizHawk.Client.EmuHawk/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using System.Diagnostics;
 using System.IO;
 using System.Collections.Generic;
@@ -15,7 +14,7 @@ using BizHawk.Client.Common;
 
 namespace BizHawk.Client.EmuHawk
 {
-	static class Program
+	internal static class Program
 	{
 		static Program()
 		{
@@ -37,27 +36,24 @@ namespace BizHawk.Client.EmuHawk
 				var vc2012 = libLoader.LoadPlatformSpecific("msvcr120.dll"); //TODO - check version?
 				var vc2010 = libLoader.LoadPlatformSpecific("msvcr100.dll"); //TODO - check version?
 				var vc2010p = libLoader.LoadPlatformSpecific("msvcp100.dll");
-				bool fail = false, warn = false;
-				warn |= d3dx9 == IntPtr.Zero;
-				fail |= vc2015 == IntPtr.Zero;
-				fail |= vc2010 == IntPtr.Zero;
-				fail |= vc2012 == IntPtr.Zero;
-				fail |= vc2010p == IntPtr.Zero;
+				var fail = vc2015 == IntPtr.Zero || vc2010 == IntPtr.Zero || vc2012 == IntPtr.Zero || vc2010p == IntPtr.Zero;
+				var warn = d3dx9 == IntPtr.Zero;
 				if (fail || warn)
 				{
-					var sw = new System.IO.StringWriter();
-					sw.WriteLine("[ OK ] .Net 4.6.1 (You couldn't even get here without it)");
-					sw.WriteLine("[{0}] Direct3d 9", d3dx9 == IntPtr.Zero ? "FAIL" : " OK ");
-					sw.WriteLine("[{0}] Visual C++ 2010 SP1 Runtime", (vc2010 == IntPtr.Zero || vc2010p == IntPtr.Zero) ? "FAIL" : " OK ");
-					sw.WriteLine("[{0}] Visual C++ 2012 Runtime", (vc2012 == IntPtr.Zero) ? "FAIL" : " OK ");
-					sw.WriteLine("[{0}] Visual C++ 2015 Runtime", (vc2015 == IntPtr.Zero) ? "FAIL" : " OK ");
-					var str = sw.ToString();
-					var box = new BizHawk.Client.EmuHawk.CustomControls.PrereqsAlert(!fail);
-					box.textBox1.Text = str;
+					var alertLines = new[]
+					{
+						"[ OK ] .NET CLR (You wouldn't even get here without it)",
+						$"[{(d3dx9 == IntPtr.Zero ? "WARN" : " OK ")}] Direct3d 9",
+						$"[{(vc2010 == IntPtr.Zero || vc2010p == IntPtr.Zero ? "FAIL" : " OK ")}] Visual C++ 2010 SP1 Runtime",
+						$"[{(vc2012 == IntPtr.Zero ? "FAIL" : " OK ")}] Visual C++ 2012 Runtime",
+						$"[{(vc2015 == IntPtr.Zero ? "FAIL" : " OK ")}] Visual C++ 2015 Runtime"
+					};
+					var box = new BizHawk.Client.EmuHawk.CustomControls.PrereqsAlert(!fail)
+					{
+						textBox1 = { Text = string.Concat("\n", alertLines) }
+					};
 					box.ShowDialog();
-					if (!fail) { }
-					else
-						System.Diagnostics.Process.GetCurrentProcess().Kill();
+					if (fail) System.Diagnostics.Process.GetCurrentProcess().Kill();
 				}
 
 				libLoader.FreePlatformSpecific(d3dx9);
@@ -67,11 +63,11 @@ namespace BizHawk.Client.EmuHawk
 				libLoader.FreePlatformSpecific(vc2010p);
 
 				// this will look in subdirectory "dll" to load pinvoked stuff
-				string dllDir = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "dll");
+				var dllDir = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "dll");
 				SetDllDirectory(dllDir);
 
 				//in case assembly resolution fails, such as if we moved them into the dll subdiretory, this event handler can reroute to them
-				AppDomain.CurrentDomain.AssemblyResolve += new ResolveEventHandler(CurrentDomain_AssemblyResolve);
+				AppDomain.CurrentDomain.AssemblyResolve += CurrentDomain_AssemblyResolve;
 
 				//but before we even try doing that, whack the MOTW from everything in that directory (thats a dll)
 				//otherwise, some people will have crashes at boot-up due to .net security disliking MOTW.
@@ -83,14 +79,14 @@ namespace BizHawk.Client.EmuHawk
 		}
 
 		[STAThread]
-		static int Main(string[] args)
+		private static int Main(string[] args)
 		{
 			return SubMain(args);
 		}
 
 		//NoInlining should keep this code from getting jammed into Main() which would create dependencies on types which havent been setup by the resolver yet... or something like that
 		[System.Runtime.CompilerServices.MethodImpl(System.Runtime.CompilerServices.MethodImplOptions.NoInlining)]
-		static int SubMain(string[] args)
+		private static int SubMain(string[] args)
 		{
 			// this check has to be done VERY early.  i stepped through a debug build with wrong .dll versions purposely used,
 			// and there was a TypeLoadException before the first line of SubMain was reached (some static ColorType init?)
@@ -112,7 +108,7 @@ namespace BizHawk.Client.EmuHawk
 
 			HawkFile.ArchiveHandlerFactory = new SevenZipSharpArchiveHandler();
 
-			ArgParser argParser = new ArgParser();
+			var argParser = new ArgParser();
 			argParser.ParseArguments(args);
 			if (argParser.cmdConfigFile != null) PathManager.SetDefaultIniPath(argParser.cmdConfigFile);
 
@@ -132,13 +128,9 @@ namespace BizHawk.Client.EmuHawk
 			BizHawk.Client.Common.StringLogUtil.DefaultToAWE = Global.Config.MoviesInAWE;
 
 			// super hacky! this needs to be done first. still not worth the trouble to make this system fully proper
-			for (int i = 0; i < args.Length; i++)
+			if (Array.Exists(args, arg => arg.StartsWith("--gdi", StringComparison.InvariantCultureIgnoreCase)))
 			{
-				var arg = args[i].ToLower();
-				if (arg.StartsWith("--gdi"))
-				{
-					Global.Config.DispMethod = Config.EDispMethod.GdiPlus;
-				}
+				Global.Config.DispMethod = Config.EDispMethod.GdiPlus;
 			}
 
 			// create IGL context. we do this whether or not the user has selected OpenGL, so that we can run opengl-based emulator cores
@@ -161,8 +153,7 @@ namespace BizHawk.Client.EmuHawk
 				}
 				catch(Exception ex)
 				{
-					var e2 = new Exception("Initialization of Direct3d 9 Display Method failed; falling back to GDI+", ex);
-					new ExceptionBox(e2).ShowDialog();
+					new ExceptionBox(new Exception("Initialization of Direct3d 9 Display Method failed; falling back to GDI+", ex)).ShowDialog();
 
 					// fallback
 					Global.Config.DispMethod = Config.EDispMethod.GdiPlus;
@@ -174,8 +165,7 @@ namespace BizHawk.Client.EmuHawk
 				GlobalWin.GL = GlobalWin.IGL_GL;
 
 				// check the opengl version and dont even try to boot this crap up if its too old
-				int version = GlobalWin.IGL_GL.Version;
-				if (version < 200)
+				if (GlobalWin.IGL_GL.Version < 200)
 				{
 					// fallback
 					Global.Config.DispMethod = Config.EDispMethod.GdiPlus;
@@ -190,8 +180,8 @@ namespace BizHawk.Client.EmuHawk
 			}
 			catch(Exception ex)
 			{
-				var e2 = new Exception("Initialization of Display Method failed; falling back to GDI+", ex);
-				new ExceptionBox(e2).ShowDialog();
+				new ExceptionBox(new Exception("Initialization of Display Method failed; falling back to GDI+", ex)).ShowDialog();
+
 				//fallback
 				Global.Config.DispMethod = Config.EDispMethod.GdiPlus;
 				goto REDO_DISPMETHOD;
@@ -204,7 +194,7 @@ namespace BizHawk.Client.EmuHawk
 				//It isn't clear whether we need the earlier SetDllDirectory(), but I think we do.
 				//note: this is pasted instead of being put in a static method due to this initialization code being sensitive to things like that, and not wanting to cause it to break
 				//pasting should be safe (not affecting the jit order of things)
-				string dllDir = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "dll");
+				var dllDir = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "dll");
 				SetDllDirectory(dllDir);
 			}
 
@@ -214,7 +204,7 @@ namespace BizHawk.Client.EmuHawk
 				{
 					try
 					{
-						new SingleInstanceController(args).Run(args);
+						new SingleInstanceController(args).Run();
 					}
 					catch (ObjectDisposedException)
 					{
@@ -273,20 +263,21 @@ namespace BizHawk.Client.EmuHawk
 			return GlobalWin.ExitCode;
 		} //SubMain
 
-		//declared here instead of a more usual place to avoid dependencies on the more usual place
 #if WINDOWS
+		//declared here instead of a more usual place to avoid dependencies on the more usual place
+
 		[DllImport("kernel32.dll", SetLastError = true)]
-		static extern uint SetDllDirectory(string lpPathName);
+		private static extern uint SetDllDirectory(string lpPathName);
 
 		[DllImport("kernel32.dll", EntryPoint = "DeleteFileW", SetLastError = true, CharSet = CharSet.Unicode, ExactSpelling = true)]
-		static extern bool DeleteFileW([MarshalAs(UnmanagedType.LPWStr)]string lpFileName);
+		private static extern bool DeleteFileW([MarshalAs(UnmanagedType.LPWStr)]string lpFileName);
 
-		public static void RemoveMOTW(string path)
+		private static void RemoveMOTW(string path)
 		{
 			DeleteFileW($"{path}:Zone.Identifier");
 		}
 
-		static void WhackAllMOTW(string dllDir)
+		private static void WhackAllMOTW(string dllDir)
 		{
 			var todo = new Queue<DirectoryInfo>(new[] { new DirectoryInfo(dllDir) });
 			while (todo.Count > 0)
@@ -298,14 +289,12 @@ namespace BizHawk.Client.EmuHawk
 				foreach (var fi in di.GetFiles("*.exe"))
 					RemoveMOTW(fi.FullName);
 			}
-
 		}
 #endif
 
-
-		static Assembly CurrentDomain_AssemblyResolve(object sender, ResolveEventArgs args)
+		private static Assembly CurrentDomain_AssemblyResolve(object sender, ResolveEventArgs args)
 		{
-			string requested = args.Name;
+			var requested = args.Name;
 
 			//mutate filename depending on selection of lua core. here's how it works
 			//1. we build NLua to the output/dll/lua directory. that brings KopiLua with it
@@ -313,52 +302,43 @@ namespace BizHawk.Client.EmuHawk
 			//3. When NLua assembly attempts to load, it can't find it
 			//I. if LuaInterface is selected by the user, we switch to requesting that.
 			//     (those DLLs are built into the output/DLL directory)
-			//II. if NLua is selected by the user, we skip over this part; 
+			//II. if NLua is selected by the user, we skip over this part;
 			//    later, we look for NLua or KopiLua assembly names and redirect them to files located in the output/DLL/nlua directory
 			if (new AssemblyName(requested).Name == "NLua")
 			{
 				//this method referencing Global.Config makes assemblies get loaded, which isnt smart from the assembly resolver.
 				//so.. we're going to resort to something really bad.
 				//avert your eyes.
-				bool UseNLua = true;
-				string configPath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "config.ini");
-				if (File.Exists(configPath))
+				var configPath = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "config.ini");
+				if (File.Exists(configPath)
+					&& (Array.Find(File.ReadAllLines(configPath), line => line.Contains("  \"UseNLua\": ")) ?? string.Empty)
+						.Contains("false"))
 				{
-					var cfg = File.ReadAllLines(configPath);
-					var usenlua_key = cfg.FirstOrDefault(line=>line.Contains("  \"UseNLua\": "));
-					if (usenlua_key != null)
-						if (usenlua_key.Contains("false"))
-							UseNLua = false;
+					requested = "LuaInterface";
 				}
-				
-				if (UseNLua) { }
-				else requested = "LuaInterface";
 			}
 
 			lock (AppDomain.CurrentDomain)
 			{
-				var asms = AppDomain.CurrentDomain.GetAssemblies();
-				foreach (var asm in asms)
-					if (asm.FullName == requested)
-						return asm;
+				var firstAsm = Array.Find(AppDomain.CurrentDomain.GetAssemblies(), asm => asm.FullName == requested);
+				if (firstAsm != null) return firstAsm;
 
 				//load missing assemblies by trying to find them in the dll directory
-				string dllname = $"{new AssemblyName(requested).Name}.dll";
-				string directory = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "dll");
-				string simpleName = new AssemblyName(requested).Name;
+				var dllname = $"{new AssemblyName(requested).Name}.dll";
+				var directory = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), "dll");
+				var simpleName = new AssemblyName(requested).Name;
 				if (simpleName == "NLua" || simpleName == "KopiLua") directory = Path.Combine(directory, "nlua");
-				string fname = Path.Combine(directory, dllname);
-				if (!File.Exists(fname)) return null;
-
-				//it is important that we use LoadFile here and not load from a byte array; otherwise mixed (managed/unamanged) assemblies can't load
-				return Assembly.LoadFile(fname);
+				var fname = Path.Combine(directory, dllname);
+				//it is important that we use LoadFile here and not load from a byte array; otherwise mixed (managed/unmanaged) assemblies can't load
+				return File.Exists(fname) ? Assembly.LoadFile(fname) : null;
 			}
 		}
 
 #if WINDOWS
-		public class SingleInstanceController : WindowsFormsApplicationBase
+		private class SingleInstanceController : WindowsFormsApplicationBase
 		{
-			readonly string[] cmdArgs;
+			private readonly string[] cmdArgs;
+
 			public SingleInstanceController(string[] args)
 			{
 				cmdArgs = args;
@@ -366,10 +346,12 @@ namespace BizHawk.Client.EmuHawk
 				StartupNextInstance += this_StartupNextInstance;
 			}
 
-			void this_StartupNextInstance(object sender, StartupNextInstanceEventArgs e)
+			public void Run() => Run(cmdArgs);
+
+			private void this_StartupNextInstance(object sender, StartupNextInstanceEventArgs e)
 			{
 				if (e.CommandLine.Count >= 1)
-					(MainForm as MainForm).LoadRom(e.CommandLine[0], new MainForm.LoadRomArgs() { OpenAdvanced = new OpenAdvanced_OpenRom() });
+					((MainForm)MainForm).LoadRom(e.CommandLine[0], new MainForm.LoadRomArgs { OpenAdvanced = new OpenAdvanced_OpenRom() });
 			}
 
 			protected override void OnCreateMainForm()
@@ -378,11 +360,9 @@ namespace BizHawk.Client.EmuHawk
 				var title = MainForm.Text;
 				MainForm.Show();
 				MainForm.Text = title;
-				GlobalWin.ExitCode = (MainForm as MainForm).ProgramRunLoop();
+				GlobalWin.ExitCode = ((MainForm)MainForm).ProgramRunLoop();
 			}
 		}
-
-
 #endif
 	}
 }
